chore: изменен тип поля Boost.Value

+++ models/boost.go
@@ -5,6 +5,6 @@ type BoostType string
 type Boost struct {
 	ID			uint	 	`json:"id" gorm:"primary_key"`
 	BoostType	BoostType	`json:"boost_type" gorm:"type:boost_type"`
-	Value       uint		`json:"value"`
+	Value       float32		`json:"value"`
 	UpgradeID   uint		`json:"-"`
 }
